tools/gdocs-style-extract/pyproject.toml
[project]
name = "gdocs-style-extract"
version = "0.1.0"
description = "Extract canonical styling information from Google Docs via the Docs REST API as JSON plus Markdown summaries."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "ShruggieTech" },
]
dependencies = [
    "google-api-python-client>=2.120.0",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
]

[project.scripts]
gdocs-style-extract = "gdocs_style_extract.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gdocs_style_extract"]
